Remove scrollback from alt screen buffer
To reduce memory usage the alt screen buffer is now initialized with a
history size of 0, this should have no impact on the primary screen
buffer but reduce memory usage.

Shift + Scrolling is still sent to alacritty as a scroll event, but
nothing happens because the scrollback buffer size is 0, it might be
worth changing this but it seems in line with other applications.