Refactor automatic builds
A few changes have been made to the automatic building of Alacritty
binaries with Travis.

Since providing binaries on Linux isn't easily possible because of
dynamic linking, I've decided to leave this up to package maintainers.
Since the .deb format allows us to target a relatively large audience
with hopefully consistent library locations, this is the only Linux
binary which will be provided.

The main travis sections (install/script/before_deploy) have also been
moved out of the .travis.yml to make it easier to read, maintain and
extend the different steps of the CI process.

Since checking for the Rust version in CI is enough to know if clippy
should be used or not, the environment variable `CLIPPY` has also been
removed, which further allowed simplifying the CI process.

Besides the executables, some auxillary files are now also published as
part of a release when they have changed since the last tagged Alacritty
release. This should make it clear for returning users when a new
version of a specific auxillary file is required.