Fix build timeout
The maximum time of 50 minutes is exceeded when Alacritty tries to build
for all Linux platforms, this is because it was effectively compiling
Alacritty from scratch four times.

By making use of the previous build artifacts, it's possible to cut this
down to two compiles using the `--no-build` option of `cargo-deb`.