Disaster strikes
Using indexing anchored at the bottom seems to just cause problems
everywhere. Both terminal and rendering start rendering at the top,
which causes strange behavior when the cursor is offset from the bottom.
Resizing a window for example shouldn't sticky the cursor to the bottom
but instead keep it at the top.

Maybe an isize with history being negative is the way to go?